requirements.txt
django
djangorestframework
djangorestframework_simplejwt
django-cors-headers
tensorflow
keras
medmnist
numpy